+++ lib/model/parameter/ContainsAdditives.dart
@@ -1,17 +1,20 @@
 import 'package:openfoodfacts/interface/Parameter.dart';
 
+// TODO: deprecated from 2021-07-21 (#194); remove when old enough
+/// "Contains additives?" search API parameter
+@Deprecated('Use WithoutAdditives instead')
 class ContainsAdditives extends Parameter {
   @override
-  String getName() {
-    return 'additives';
-  }
+  String getName() => 'additives';
 
   @override
-  String getValue() {
-    return filter! ? 'without' : '';
-  }
+  String getValue() => filter ? 'without' : '';
 
-  final bool? filter;
+  // TODO: find a less misleading parameter name than "filter"
+  // it looks like the filter is:
+  // * `ContainsAdditives(filter: true)` means "without additives"
+  // * `ContainsAdditives(filter: false)` means "no additive filter at all"
+  final bool filter;
 
-  const ContainsAdditives({this.filter});
+  const ContainsAdditives({required this.filter});
 }