+++ lib/model/parameter/SearchSimple.dart
@@ -1,15 +1,13 @@
 import 'package:openfoodfacts/interface/Parameter.dart';
 
+// TODO: deprecated from 2021-07-13 (#107); remove when old enough
+@Deprecated('No use actually; just do not use anymore')
 class SearchSimple extends Parameter {
   @override
-  String getName() {
-    return 'search_type';
-  }
+  String getName() => 'search_type';
 
   @override
-  String getValue() {
-    return active ? '1' : '0';
-  }
+  String getValue() => active ? '1' : '0';
 
   final bool active;