+++ lib/model/parameter/SortBy.dart
@@ -1,34 +1,26 @@
 import 'package:openfoodfacts/interface/Parameter.dart';
 
+/// Sort search API parameter
 class SortBy extends Parameter {
-  @override
-  String getName() {
-    return 'sort_by';
-  }
+  static const Map<SortOption, String> _VALUES = {
+    SortOption.PRODUCT_NAME: 'product_name',
+    SortOption.CREATED: 'created_t',
+    SortOption.EDIT: 'last_modified_t',
+    SortOption.POPULARITY: 'unique_scans_n',
+  };
 
   @override
-  String getValue() {
-    switch (option) {
-      case SortOption.PRODUCT_NAME:
-        return 'product_name';
+  String getName() => 'sort_by';
 
-      case SortOption.CREATED:
-        return 'created_t';
-
-      case SortOption.EDIT:
-        return 'last_modified_t';
-
-      case SortOption.POPULARITY:
-      default:
-        return 'unique_scans_n';
-    }
-  }
+  @override
+  String getValue() => _VALUES[option] ?? 'unique_scans_n';
 
   final SortOption? option;
 
   const SortBy({this.option});
 }
 
+/// Possible sort options for search API
 enum SortOption {
   POPULARITY,
   PRODUCT_NAME,