+++ lib/openfoodfacts.dart
@@ -5,10 +5,13 @@ import 'dart:async';
 
 import 'package:http/http.dart';
 import 'package:openfoodfacts/model/OcrIngredientsResult.dart';
+import 'package:openfoodfacts/utils/AbstractQueryConfiguration.dart';
 import 'package:openfoodfacts/utils/OcrField.dart';
 import 'package:openfoodfacts/utils/PnnsGroupQueryConfiguration.dart';
 import 'package:openfoodfacts/utils/PnnsGroups.dart';
+import 'package:openfoodfacts/utils/ProductListQueryConfiguration.dart';
 import 'package:openfoodfacts/utils/QueryType.dart';
+import 'package:openfoodfacts/utils/TagType.dart';
 
 import 'model/Insight.dart';
 import 'model/RobotoffQuestion.dart';
@@ -20,7 +23,6 @@ import 'model/SpellingCorrections.dart';
 import 'model/Status.dart';
 import 'model/User.dart';
 
-import 'model/parameter/OutputFormat.dart';
 import 'utils/HttpHelper.dart';
 import 'utils/LanguageHelper.dart';
 import 'utils/ProductHelper.dart';
@@ -69,6 +71,9 @@ class OpenFoodAPIClient {
   /// Add the given product to the database.
   /// By default the query will hit the PROD DB
   /// Returns a Status object as result.
+  ///
+  /// Please read the language mechanics explanation if you intend to display
+  /// or update data in specific language: https://github.com/openfoodfacts/openfoodfacts-dart/blob/master/DOCUMENTATION.md#about-languages-mechanics
   static Future<Status> saveProduct(User user, Product product,
       {QueryType queryType = QueryType.PROD}) async {
     var parameterMap = <String, String>{};
@@ -79,8 +84,19 @@ class OpenFoodAPIClient {
         scheme: URI_SCHEME,
         host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_TEST_HOST,
         path: '/cgi/product_jqm2.pl');
-
-    Response response = await HttpHelper()
+    if (product.nutriments != null) {
+      final Map<String, String> rawNutrients = product.nutriments!.toData();
+      for (final MapEntry<String, String> entry in rawNutrients.entries) {
+        String key = 'nutriment_${entry.key}';
+        final int pos = key.indexOf('_100g');
+        if (pos != -1) {
+          key = key.substring(0, pos);
+        }
+        parameterMap[key] = entry.value;
+      }
+    }
+    parameterMap.remove('nutriments');
+    final Response response = await HttpHelper()
         .doPostRequest(productUri, parameterMap, user, queryType: queryType);
     var status = Status.fromJson(json.decode(response.body));
     return status;
@@ -105,9 +121,8 @@ class OpenFoodAPIClient {
         host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_TEST_HOST,
         path: '/cgi/product_image_upload.pl');
 
-    return await HttpHelper().doMultipartRequest(
-        imageUri, dataMap, fileMap, user,
-        queryType: queryType);
+    return await HttpHelper().doMultipartRequest(imageUri, dataMap,
+        files: fileMap, user: user, queryType: queryType);
   }
 
   /// Returns the product for the given barcode.
@@ -135,6 +150,9 @@ class OpenFoodAPIClient {
   /// The ProductResult does not contain a product, if the product is not available.
   /// ingredients, images and product name will be prepared for the given language.
   /// By default the query will hit the PROD DB
+  ///
+  /// Please read the language mechanics explanation if you intend to show
+  /// or update data in specific language: https://github.com/openfoodfacts/openfoodfacts-dart/blob/master/DOCUMENTATION.md#about-languages-mechanics
   static Future<ProductResult> getProduct(
       ProductQueryConfiguration configuration,
       {User? user,
@@ -153,8 +171,6 @@ class OpenFoodAPIClient {
     if (result.product != null) {
       ProductHelper.removeImages(result.product!, configuration.language);
       ProductHelper.createImageUrls(result.product!, queryType: queryType);
-      final translatedLang = configuration.lc ?? configuration.language?.code;
-      result.product!.translatedLang = LanguageHelper.fromJson(translatedLang);
     }
 
     return result;
@@ -171,9 +187,7 @@ class OpenFoodAPIClient {
   static Future<SearchResult> searchProducts(
       User? user, ProductSearchQueryConfiguration configuration,
       {QueryType queryType = QueryType.PROD}) async {
-    const outputFormat = OutputFormat(format: Format.JSON);
     var queryParameters = configuration.getParametersMap();
-    queryParameters[outputFormat.getName()] = outputFormat.getValue();
 
     var searchUri = Uri(
         scheme: URI_SCHEME,
@@ -186,20 +200,41 @@ class OpenFoodAPIClient {
     final jsonStr = _replaceQuotes(response.body);
     var result = SearchResult.fromJson(json.decode(jsonStr));
 
-    result.products!.asMap().forEach((index, product) {
-      ProductHelper.removeImages(product, configuration.language);
-    });
+    _removeImages(result, configuration);
 
     return result;
   }
 
-  /// By default the query will hit the PROD DB
+  /// Search the OpenFoodFacts product database: multiple barcodes in input.
+  static Future<SearchResult> getProductList(
+    User? user,
+    ProductListQueryConfiguration configuration, {
+    QueryType queryType = QueryType.PROD,
+  }) async {
+    final Uri uri = Uri(
+        scheme: URI_SCHEME,
+        host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_TEST_HOST,
+        path: 'products/${configuration.barcodes.join(',')}.json',
+        queryParameters: configuration.getParametersMap());
+
+    final Response response =
+        await HttpHelper().doGetRequest(uri, user: user, queryType: queryType);
+
+    final String jsonStr = _replaceQuotes(response.body);
+    final SearchResult result = SearchResult.fromJson(json.decode(jsonStr));
+
+    _removeImages(result, configuration);
+
+    return result;
+  }
+
+  // TODO: deprecated from 2021-07-13 (#92); remove when old enough
+  @Deprecated(
+      'Use PnnsGroup2Filter with ProductSearchQueryConfiguration instead')
   static Future<SearchResult> queryPnnsGroup(
       User user, PnnsGroupQueryConfiguration configuration,
       {QueryType queryType = QueryType.PROD}) async {
-    const outputFormat = OutputFormat(format: Format.JSON);
     var queryParameters = configuration.getParametersMap();
-    queryParameters[outputFormat.getName()] = outputFormat.getValue();
 
     var searchUri = Uri(
         scheme: URI_SCHEME,
@@ -212,11 +247,20 @@ class OpenFoodAPIClient {
     final jsonStr = _replaceQuotes(response.body);
     var result = SearchResult.fromJson(json.decode(jsonStr));
 
-    result.products!.asMap().forEach((index, product) {
-      ProductHelper.removeImages(product, configuration.language);
-    });
+    _removeImages(result, configuration);
 
     return result;
+  }
+
+  static void _removeImages(
+    final SearchResult searchResult,
+    final AbstractQueryConfiguration configuration,
+  ) {
+    if (searchResult.products != null) {
+      searchResult.products!.asMap().forEach((index, product) {
+        ProductHelper.removeImages(product, configuration.language);
+      });
+    }
   }
 
   /// By default the query will hit the PROD DB
@@ -411,12 +455,11 @@ class OpenFoodAPIClient {
   /// Returns the ingredients using OCR.
   /// By default the query will use the Google Cloud Vision.
   /// By default the query will hit the PROD DB
-
   static Future<OcrIngredientsResult> extractIngredients(
       User user, String barcode, OpenFoodFactsLanguage language,
       {OcrField ocrField = OcrField.GOOGLE_CLOUD_VISION,
       QueryType queryType = QueryType.PROD}) async {
-    var productUri = Uri(
+    var ocrUri = Uri(
         scheme: URI_SCHEME,
         host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_PROD_HOST,
         path: '/cgi/ingredients.pl',
@@ -428,18 +471,142 @@ class OpenFoodAPIClient {
         });
 
     Response response = await HttpHelper()
-        .doGetRequest(productUri, user: user, queryType: queryType);
+        .doGetRequest(ocrUri, user: user, queryType: queryType);
 
     OcrIngredientsResult result = OcrIngredientsResult.fromJson(
         json.decode(utf8.decode(response.bodyBytes)));
     return result;
   }
 
-  /// login on the main page - not used
-  static Future<String> login(User user) async {
-    var loginUri = Uri(scheme: URI_SCHEME, host: URI_PROD_HOST);
+  /// Give user suggestion based on autocompleted outputs
+  /// The expected output language can be set otherwise English will be used by default
+  /// The TagType is required
+  /// Returns a List of suggestions
+  /// By default the query will hit the PROD DB
+  static Future<List<dynamic>> getAutocompletedSuggestions(TagType tagType,
+      {String input = '',
+      OpenFoodFactsLanguage language = OpenFoodFactsLanguage.ENGLISH,
+      QueryType queryType = QueryType.PROD}) async {
+    var suggestionUri = Uri(
+        scheme: URI_SCHEME,
+        host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_PROD_HOST,
+        path: '/cgi/suggest.pl',
+        queryParameters: {
+          'tagtype': tagType.key,
+          'term': input,
+          'lc': language.code,
+        });
+
+    Response response =
+        await HttpHelper().doGetRequest(suggestionUri, queryType: queryType);
+
+    return json.decode(response.body);
+  }
+
+  /// Uses the auth.pl API to see if login was successful
+  /// Returns a bool if the login data of the provided user is correct
+  static Future<bool> login(User user,
+      {QueryType queryType = QueryType.PROD}) async {
+    var loginUri = Uri(
+      scheme: URI_SCHEME,
+      host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_TEST_HOST,
+      path: '/cgi/auth.pl',
+    );
     Response response =
         await HttpHelper().doPostRequest(loginUri, user.toData(), user);
-    return response.body;
+    return response.statusCode == 200;
+  }
+
+  /// Creates a new user
+  /// Returns [Status.status] 201 = complete; 400 = wrong inputs + [Status.error]; 500 = server error;
+  ///
+  /// When creating a [producer account](https://world.pro.openfoodfacts.org/) use [requested_org] to name the Producer or brand
+  static Future<Status> register({
+    required User user,
+    required String name,
+    required String email,
+    String? requested_org,
+    bool newsletter = true,
+    QueryType queryType = QueryType.PROD,
+  }) async {
+    var registerUri = Uri(
+      scheme: URI_SCHEME,
+      host: queryType == QueryType.PROD ? URI_PROD_HOST : URI_TEST_HOST,
+      path: '/cgi/user.pl',
+    );
+
+    Map<String, String> data = <String, String>{
+      'name': name,
+      'email': email,
+      'userid': user.userId,
+      'password': user.password,
+      'confirm_password': user.password,
+      if (requested_org != null) 'pro': 'on',
+      'pro_checkbox': '1',
+      'requested_org': requested_org ?? ' ',
+      if (newsletter) 'newsletter': 'on',
+      'action': 'process',
+      'type': 'add',
+      '.submit': 'Register',
+    };
+
+    Status status = await HttpHelper().doMultipartRequest(
+      registerUri,
+      data,
+      queryType: queryType,
+    );
+
+    //Since this is not a official endpoint the response code is always 200
+    //Here we check the response body for certain keyword to find out if the registration was complete
+    if (status.body == null) {
+      return Status(
+        status: 500,
+        error:
+            'No response, open an issue here: https://github.com/openfoodfacts/openfoodfacts-dart/issues/new',
+      );
+    } else if (status.body!.contains('loggedin')) {
+      return Status(status: 201);
+    } else if (status.body!
+        .contains('This username already exists, please choose another.')) {
+      return Status(
+        status: 400,
+        error: 'This username already exists, please choose another.',
+      );
+    } else if (status.body!.contains('The e-mail address is already used.')) {
+      return Status(
+        status: 400,
+        error:
+            'The e-mail address is already used by another user. Maybe you already have an account? You can reset the password of your other account.',
+      );
+    } else {
+      return Status(status: 400, error: 'Unrecognized request error');
+    }
+  }
+
+  /// Returns the Ecoscore description in HTML
+  static Future<String?> getEcoscoreHtmlDescription(
+    final String barcode,
+    final OpenFoodFactsLanguage language,
+  ) async {
+    const String FIELD = 'environment_infocard';
+    final Uri uri = Uri(
+      scheme: URI_SCHEME,
+      host: 'world-${language.code}.openfoodfacts.org',
+      path: '/api/v0/product/$barcode.json',
+      queryParameters: <String, String>{'fields': FIELD},
+    );
+    try {
+      final Response response = await HttpHelper().doGetRequest(uri);
+      if (response.statusCode != 200) {
+        return null;
+      }
+      final Map<String, dynamic> json =
+          jsonDecode(response.body) as Map<String, dynamic>;
+      final Map<String, dynamic> productData =
+          json['product'] as Map<String, dynamic>;
+      return productData[FIELD] as String?;
+    } catch (e) {
+      return null;
+    }
   }
 }